api/requirements.txt
Flask
flask-cors
tensorflow==2.14.0
numpy
Pillow
gunicorn